Add dummy view one trip window, resize windows
Resize all windows to same size for consistency.
Add a window for viewing one trip, accessible via double click on the row containing it or clicking on the row and then the view details button. Right now it's just a dummy window displaying the trip it received from the previous window via a MessageBox.